Table is done

+++ src/components/ConstructionTable/ConstructionTable.tsx
@@ -9,7 +9,7 @@ import {
 import {
   ConstructionData,
   ConstructionDataRow,
-  ConstructionDataRowWOChild,
+  ConstructionDataRowRespose,
   ConstructionRowRender,
 } from "./ConstructionTable.types";
 
@@ -26,7 +26,7 @@ const columns = [
 // где кроме обычных данных строки, также указываем parentId для передачи его при клике по кнопке создания
 // для передачи на сервер запроса
 // также считаем уровень для отрисовки дерева в графе Уровень
-function constructionDataToRender(
+function convertDataForRender(
   data: ConstructionData,
   parent: number | null = null,
   level: number = 1
@@ -57,7 +57,11 @@ function constructionDataToRender(
     result.push(currentItem);
 
     if (child.length) {
-      const children = constructionDataToRender(child, id, level + 1);
+      // Сортировка детей по количеству детей (сначала без детей)
+      const sortedChildren = child.sort((a, b) => a.child.length - b.child.length);
+      
+      // Рекурсивный вызов для детей
+      const children = convertDataForRender(sortedChildren, id, level + 1);
       result = result.concat(children);
     }
   });
@@ -65,34 +69,98 @@ function constructionDataToRender(
   return result;
 }
 
+// Функция для обновления данных
 function updateData(
   data: ConstructionData,
-  changedRows: ConstructionDataRowWOChild[]
-) {
+  changedRows: ConstructionDataRowRespose[]
+): ConstructionData {
+  
   const changedRowsMap = new Map(changedRows.map((row) => [row.id, row]));
 
   const updateRow = (row: ConstructionDataRow): ConstructionDataRow => {
     const changedRow = changedRowsMap.get(row.id);
     if (changedRow) {
-      return { ...row, ...changedRow, child: row.child.map(updateRow) };
+      return {
+        ...row,
+        ...changedRow,
+        child: row.child.map(updateRow)
+      };
     }
-    return { ...row, child: row.child.map(updateRow) };
+    return {
+      ...row,
+      child: row.child.map(updateRow)
+    };
   };
 
-  const result = data.map(updateRow);
-
-  return result;
+  return data.map(updateRow);
 }
 
 function deleteRow(data: ConstructionData, rowId: number): ConstructionData {
   const deleteRecursive = (rows: ConstructionData): ConstructionData => {
-    return rows.filter(row => row.id !== rowId).map(row => ({
-      ...row,
-      child: deleteRecursive(row.child)
-    }));
+    return rows
+      .filter((row) => row.id !== rowId)
+      .map((row) => ({
+        ...row,
+        child: deleteRecursive(row.child),
+      }));
   };
 
   return deleteRecursive(data);
+}
+
+// функция для добавления новой стоки как ребенка какого-то родителя
+function addNewRowToParent(
+  data: ConstructionData,
+  newRow: ConstructionRowRender
+): ConstructionData {
+  const {
+    id,
+    equipmentCosts,
+    estimatedProfit,
+    overheads,
+    rowName,
+    salary,
+    parentId,
+  } = newRow;
+  const newRowConstructData: ConstructionDataRow = {
+    id: id,
+    equipmentCosts: equipmentCosts,
+    estimatedProfit: estimatedProfit,
+    overheads: overheads,
+    rowName: rowName,
+    salary: salary,
+    total: 0,
+    mimExploitation: 0,
+    machineOperatorSalary: 0,
+    mainCosts: 0,
+    materials: 0,
+    supportCosts: 0,
+    child: [],
+  };
+
+  const addRowRecursive = (nodes: ConstructionData): ConstructionData => {
+    return nodes.map((node) => {
+      if (node.id === parentId) {
+        return {
+          ...node,
+          child: [...node.child, newRowConstructData],
+        };
+      } else {
+        return {
+          ...node,
+          child: addRowRecursive(node.child),
+        };
+      }
+    });
+  };
+
+  if (parentId === null) {
+    // Если parentId равен null, добавляем новую строку на верхний уровень
+    return [...data, newRowConstructData];
+  } else {
+    // Иначе, добавляем новую строку к родительскому элементу
+    return addRowRecursive(data);
+  }
 }
 
 function ConstructionTable(): ReactElement {
@@ -106,9 +174,24 @@ function ConstructionTable(): ReactElement {
     const fetchData = async () => {
       try {
         const dataFromServer = await getConstructionCostsListAPI();
-        setDataFromAPI(dataFromServer);
+
+        if (dataFromServer.length === 0) {
+          // если нет данных, то создаем пустую строку
+          handleCreate({
+            id: -1, // ID заглушка, потом получим нормальный от сервера
+            rowName: "",
+            salary: 0,
+            equipmentCosts: 0,
+            overheads: 0,
+            estimatedProfit: 0,
+            level: 1,
+            parentId: null,
+          });
+        } else {
+          setDataFromAPI(dataFromServer);
+        }
       } catch (error) {
-        console.error("Failed to fetch construction costs data", error);
+        console.error("Ошибка при получении всех строк ", error);
       }
     };
 
@@ -116,17 +199,50 @@ function ConstructionTable(): ReactElement {
   }, []);
 
   useEffect(() => {
-    setDataForRender(constructionDataToRender(dataFromAPI));
+    setDataForRender(convertDataForRender(dataFromAPI));
   }, [dataFromAPI]);
 
-  const handleCreate = async (newRow: ConstructionRowRender) => {
+  const handleCreate = (newRow: ConstructionRowRender) => {
+    setDataForRender((prev) => [...prev, newRow]);
+  };
+
+  const handleDeleteNewRow = (newRowId: number | null) => {
+    setDataForRender((prev) => prev.filter((row) => row.id !== newRowId));
+  };
+
+  const handleSubmitNewRow = async (newRow: ConstructionRowRender) => {
     try {
       const { current, changed } = await createConstructionCostAPI(newRow);
-      const updated = updateData(dataFromAPI, [current, ...changed]);
 
-      setDataFromAPI(updated);
+      let withNewRow;
+      if (newRow.parentId === null) {
+        // Если у новой строки нет родителя, добавляем ее на верхний уровень
+        withNewRow = [
+          ...dataFromAPI,
+          {
+            ...current,
+            child: [], // У новой строки нет детей
+          },
+        ];
+      } else {
+        // Иначе добавляем новую строку к родителю
+        withNewRow = addNewRowToParent(dataFromAPI, {
+          ...current,
+          parentId: newRow.parentId,
+          level: newRow.level,
+        });
+      }
+
+      if (changed.length) {
+        const updated = updateData(withNewRow, [...changed]);
+        setDataFromAPI(updated);
+        setDataForRender(convertDataForRender(updated));
+      } else {
+        setDataFromAPI(withNewRow);
+        setDataForRender(convertDataForRender(withNewRow));
+      }
     } catch (err) {
-      console.log("Ошибка во время обновления данных ", err);
+      console.log("Ошибка во время создания строки ", err);
     }
   };
 
@@ -134,30 +250,54 @@ function ConstructionTable(): ReactElement {
     try {
       const { current, changed } = await updateConstructionCostAPI(updatedData);
       const updated = updateData(dataFromAPI, [current, ...changed]);
-
       setDataFromAPI(updated);
     } catch (err) {
-      console.log("Ошибка во время обновления данных ", err);
+      console.log("Ошибка во время обновления строки ", err);
     }
   };
 
   const handleDelete = async (deletedRow: ConstructionRowRender) => {
+    if (deletedRow.id === -1) {
+      setDataForRender(dataForRender.filter((row) => row.id !== deletedRow.id));
+      return;
+    }
     try {
       const { changed } = await deleteConstructionCostAPI(deletedRow);
       let updated = deleteRow(dataFromAPI, deletedRow.id);
       if (changed) {
         updated = updateData(updated, [...changed]);
       }
-      
       setDataFromAPI(updated);
+
+      if (updated.length === 0) {
+        // если нет данных, то создаем пустую строку
+        handleCreate({
+          id: -1, // ID заглушка, потом получим нормальный от сервера
+          rowName: "",
+          salary: 0,
+          equipmentCosts: 0,
+          overheads: 0,
+          estimatedProfit: 0,
+          level: 1,
+          parentId: null,
+        });
+      }
     } catch (err) {
-      console.log("Ошибка во время обновления данных ", err);
+      console.log("Ошибка во время удаления строки ", err);
     }
-  }
+  };
 
   return (
     dataForRender && (
-      <Table columns={columns} data={dataForRender} onUpdate={handleUpdate} onDelete={handleDelete} onCreate={handleCreate}/>
+      <Table
+        columns={columns}
+        data={dataForRender}
+        onUpdate={handleUpdate}
+        onDelete={handleDelete}
+        onCreate={handleCreate}
+        onSubmit={handleSubmitNewRow}
+        onCancelNewRow={handleDeleteNewRow}
+      />
     )
   );
 }
